added status bar level converting system

+++ test/com/jbion/candyboxcheater/view/bindings/TestMultiBooleanBinding.java
@@ -2,8 +2,10 @@ package com.jbion.candyboxcheater.view.bindings;
 
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.IntegerProperty;
+import javafx.beans.property.LongProperty;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleIntegerProperty;
+import javafx.beans.property.SimpleLongProperty;
 
 import org.junit.Test;
 
@@ -11,7 +13,7 @@ public class TestMultiBooleanBinding {
 
 	@Test
 	public void testBinding() {
-		IntegerProperty intProp = new SimpleIntegerProperty(0);
+		LongProperty intProp = new SimpleLongProperty(0);
 		BooleanProperty[] boolProps = new BooleanProperty[5];
 		for (int i = 0; i < boolProps.length; i++) {
 			boolProps[i] = new SimpleBooleanProperty(false);
@@ -32,19 +34,19 @@ public class TestMultiBooleanBinding {
 		testSetBool(intProp, boolProps, 1, false);
 	}
 	
-	private static void testSetInt(IntegerProperty intProp, BooleanProperty[] boolProps, int i) {
+	private static void testSetInt(LongProperty intProp, BooleanProperty[] boolProps, int i) {
 		System.out.print("Setting int to " + i + " \t");
 		intProp.set(i);
 		print(intProp, boolProps);
 	}
 	
-	private static void testSetBool(IntegerProperty intProp, BooleanProperty[] boolProps, int i, boolean value) {
+	private static void testSetBool(LongProperty intProp, BooleanProperty[] boolProps, int i, boolean value) {
 		System.out.print("Setting bool[" + i + "] to " + value + " \t");
 		boolProps[i].set(value);
 		print(intProp, boolProps);
 	}
 
-	private static void print(IntegerProperty intProp, BooleanProperty[] boolProps) {
+	private static void print(LongProperty intProp, BooleanProperty[] boolProps) {
 		System.out.print(intProp.get() + " - ");
 		for (BooleanProperty prop : boolProps) {
 			System.out.print(prop.get() + " ");
